generic stream handling model, use new proxima stream client, handle new tokens stream
