+++ .env
@@ -4,7 +4,7 @@
 # Prisma supports the native connection string format for PostgreSQL, MySQL, SQLite, SQL Server, MongoDB (Preview) and CockroachDB (Preview).
 # See the documentation for all the connection string options: https://pris.ly/d/connection-strings
 
-DEV_SCHEMA=mangrove20
+DEV_SCHEMA=mangrove26
 DATABASE_URL_DEV=postgresql://postgres:postgres@localhost:5432/postgres?schema=${DEV_SCHEMA}
 DATABASE_URL_CLOUD="postgresql://..."
 DATABASE_URL_STAGE="postgresql://..."
