+++ package.json
@@ -28,9 +28,12 @@
   },
   "dependencies": {
     "@prisma/client": "3.11.0",
-    "@proximaone/stream-client-js": "^0.4.0",
+    "@proximaone/stream-client-js": "^0.7.0",
+    "@proximaone/stream-schema-mangrove": "^0.1.9",
+    "@proximaone/stream-schema-fungible-token": "^0.2.0",
     "@types/graphql-fields": "^1.3.4",
     "apollo-server": "^3.6.3",
+    "async-retry": "^1.3.3",
     "bignumber.js": "^9.0.2",
     "class-validator": "^0.13.2",
     "graphql": "^15.3.0",
@@ -38,7 +41,7 @@
     "graphql-scalars": "^1.14.1",
     "lodash": "^4.17.21",
     "node-fetch-cache": "^3.0.5",
-    "prisma": "~3.11.0",
+    "prisma": "~3.11.1",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.5.4",
     "tslib": "^2.3.1",
@@ -46,6 +49,7 @@
     "typegraphql-prisma": "^0.19.0"
   },
   "devDependencies": {
+    "@types/async-retry": "^1.4.3",
     "@types/jest": "^27.4.0",
     "@types/lodash": "^4.14.178",
     "@types/node": "^14.18.4",
