--- src/eventHandler/index.ts
@@ -1,3 +1,0 @@
-export * from "./eventHandler";
-export * as events from "./events";
-export * from "./model";
